requirements.txt
Django==3.2.10
matplotlib==3.3.2
Pillow==9.0.1
